+++ protected/humhub/modules/file/widgets/UploadInput.php
@@ -159,6 +159,13 @@ class UploadInput extends JsWidget
      */
     public $attach = true;
 
+    /**
+     * Pre-uploaded File GUIDs to be attached to the new content
+     *
+     * @var array
+     */
+    public $fileList = [];
+
     public function init()
     {
         parent::init();
@@ -182,6 +189,9 @@ class UploadInput extends JsWidget
             foreach (static::getSubmittedFiles($this->model, $this->attribute, $this->submitName) as $file) {
                 $result .= Html::hiddenInput($this->submitName, $file->guid);
             }
+        }
+        foreach ((array)$this->fileList as $file) {
+            $result .= Html::hiddenInput($this->submitName, $file);
         }
 
         return $result;